requirements.txt
sqlalchemy-iris
flask
feedparser
faker
requests